+++ src/components/SelectFormView/index.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
+import { Link as RouterLink } from 'react-router-dom';
 
 // material-ui
-import { Grid, Button, ButtonGroup, CardActions, CardContent, TextField, Typography } from '@mui/material';
+import { Grid, Button, CardContent, TextField, Typography } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import MainCard from 'components/MainCard';
 
@@ -9,16 +10,19 @@ import MainCard from 'components/MainCard';
 import { Edit, Delete, AddToPhotos } from '@mui/icons-material';
 
 // redux
-import { createForm, activeStep } from 'store/reducers/form';
+import { createForm, activeStep, removeForm } from 'store/reducers/form';
 import { useDispatch, useSelector } from 'react-redux';
 
 const SelectFormView = () => {
   const { forms } = useSelector((state) => state.form);
   const dispatch = useDispatch();
-  const goToForm = (formId) => dispatch(activeStep(formId));
   const theme = useTheme();
   const [addNewTitle, setAddNewTitle] = useState('');
-  const addForm = () => dispatch(createForm({ userTitle: addNewTitle }));
+  const addForm = () => {
+    setAddNewTitle('');
+    dispatch(createForm({ userTitle: addNewTitle }));
+  };
+  const deleteForm = (formId) => dispatch(removeForm(formId));
 
   const formCardsDom = () => {
     const formIds = Object.keys(forms);
@@ -28,11 +32,9 @@ const SelectFormView = () => {
         console.log('formData', formData);
 
         return (
-          <Grid key={formId} item xs={4}>
+          <Grid key={formId} item xs={12} sm={6} md={4}>
             <MainCard
               sx={{
-                backgroundColor: theme.palette.grey[700],
-                color: theme.palette.secondary.A100,
                 height: '100%'
               }}
               content={false}
@@ -41,14 +43,18 @@ const SelectFormView = () => {
                 <Typography variant="h3" paragraph>
                   {formData.userTitle || 'Formular: ' + formId}
                 </Typography>
-                <ButtonGroup>
-                  <Button color="light" startIcon={<Edit />} variant="contained" onClick={() => goToForm(formId)}>
-                    Bearbeiten
-                  </Button>
-                  <Button sx={{ height: '100%' }} color="light" variant="contained" onClick={() => goToForm(formId)}>
-                    <Delete />
-                  </Button>
-                </ButtonGroup>
+                <Grid container spacing={1}>
+                  <Grid item>
+                    <Button component={RouterLink} variant="contained" startIcon={<Edit />} to={`/form/${formId}`}>
+                      Bearbeiten
+                    </Button>
+                  </Grid>
+                  <Grid item>
+                    <Button color="secondary" variant="contained" onClick={() => deleteForm(formId)}>
+                      <Delete />
+                    </Button>
+                  </Grid>
+                </Grid>
               </CardContent>
             </MainCard>
           </Grid>
@@ -59,9 +65,7 @@ const SelectFormView = () => {
       <>
         <Grid container spacing={2} sx={{ marginBottom: theme.spacing(2) }}>
           {formCards}
-        </Grid>
-        <Grid container spacing={2}>
-          <Grid item xs={4}>
+          <Grid item xs={12} sm={6} md={8}>
             <MainCard key={'newForm'} sx={{ height: '100%' }} content={false}>
               <CardContent>
                 <Typography variant="h3" paragraph>
@@ -70,7 +74,12 @@ const SelectFormView = () => {
                 <Typography paragraph>Erstellen Sie ein neues Formular</Typography>
                 <Grid container spacing={1}>
                   <Grid item xs>
-                    <TextField sx={{ width: '100%' }} placeholder="Name des Formulars" onChange={(e) => setAddNewTitle(e.target.value)} />
+                    <TextField
+                      value={addNewTitle}
+                      sx={{ width: '100%' }}
+                      placeholder="Name des Formulars"
+                      onChange={(e) => setAddNewTitle(e.target.value)}
+                    />
                   </Grid>
                   <Grid item xs="auto">
                     <Button
