+++ src/routes/MainRoutes.js
@@ -9,6 +9,7 @@ const DashboardDefault = Loadable(lazy(() => import('pages/dashboard')));
 
 // render - sample page
 const SamplePage = Loadable(lazy(() => import('pages/extra-pages/SamplePage')));
+const Form = Loadable(lazy(() => import('pages/extra-pages/Form')));
 
 // render - utilities
 const Typography = Loadable(lazy(() => import('pages/components-overview/Typography')));
@@ -25,6 +26,10 @@ const MainRoutes = {
     {
       path: '/',
       element: <SamplePage />
+    },
+    {
+      path: 'form/:userId',
+      element: <Form />
     },
     {
       path: 'color',
