+++ src/themes/overrides/FormControl.js
@@ -4,9 +4,7 @@ export default function MuiFormControl() {
   return {
     MuiFormControl: {
       styleOverrides: {
-        root: {
-          // marginBottom: theme.spacing(2)
-        }
+        root: {}
       }
     }
   };
