+++ src/themes/overrides/FormHelperText.js
@@ -1,0 +1,14 @@
+// ==============================|| OVERRIDES - INPUT LABEL ||============================== //
+
+export default function FormHelperText() {
+  return {
+    MuiFormHelperText: {
+      styleOverrides: {
+        root: {
+          marginLeft: '0',
+          marginRight: '0'
+        }
+      }
+    }
+  };
+}
