+++ src/themes/overrides/TextField.js
@@ -1,0 +1,11 @@
+// ==============================|| OVERRIDES - INPUT LABEL ||============================== //
+
+export default function TextField() {
+  return {
+    MuiTextField: {
+      defaultProps: {
+        InputLabelProps: { shrink: false }
+      }
+    }
+  };
+}
