+++ src/themes/overrides/index.js
@@ -9,7 +9,8 @@ import Checkbox from './Checkbox';
 import Chip from './Chip';
 import IconButton from './IconButton';
 import InputLabel from './InputLabel';
-import MuiFormControl from './MuiFormControl';
+import FormControl from './FormControl';
+import TextField from './TextField';
 import MuiCard from './MuiCard';
 import LinearProgress from './LinearProgress';
 import Link from './Link';
@@ -19,6 +20,7 @@ import Tab from './Tab';
 import TableCell from './TableCell';
 import Tabs from './Tabs';
 import Typography from './Typography';
+import FormHelperText from './FormHelperText';
 
 // ==============================|| OVERRIDES - MAIN ||============================== //
 
@@ -30,7 +32,9 @@ export default function ComponentsOverrides(theme) {
     Checkbox(theme),
     Chip(theme),
     IconButton(theme),
-    MuiFormControl(),
+    FormControl(),
+    TextField(),
+    FormHelperText(),
     MuiCard(),
     InputLabel(theme),
     LinearProgress(),
